Set hotplugging mode automatically
If the governor ends in 'x', we need to set hotplugging to 1 so our CPU isn't stuck at maxfreq... if the governor doesn't end in 'x', then we don't deep sleep when hotplugging is 1, so it needs to be 0.

I made vanirtweakapply do this automatically by checking to see if scaleGov ends in x with a regex. Setting hotplug wrong causes major problems, and the right setting changes depending on something people may care about playing with... so we should handle it for them.

Change-Id: Ib15d47d4c627af943e2bfce21014bae2124e07a1