Edited for better looks
Animated the SMS icon--because you can't mess up a theme with a theme
edited the colors of the sms app, better looking--still more to do
swapped out the smileys in the sms app--look MUCH better
Uh. More blackness I guess. Will be adding in more as time goes on,
for now, this can be merged.

Changed the preview png.

Change-Id: Ic5b52591dbadfdc4bd4934b22ced7a115f004382